+++ src/com/kaylerrenslow/armaDialogCreator/data/export/ProjectExporter.java
@@ -11,33 +11,24 @@ import com.kaylerrenslow.armaDialogCreator.control.PropertyType;
 import com.kaylerrenslow.armaDialogCreator.control.sv.SerializableValue;
 import com.kaylerrenslow.armaDialogCreator.data.Project;
 import com.kaylerrenslow.armaDialogCreator.main.Lang;
+import com.kaylerrenslow.armaDialogCreator.util.IndentedStringBuilder;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
-import java.io.*;
+import java.io.File;
+import java.io.FileOutputStream;
+import java.io.IOException;
+import java.io.OutputStream;
 import java.util.ResourceBundle;
+import java.util.function.Function;
 
 /**
  Created by Kayler on 09/13/2016.
  */
 public class ProjectExporter {
-	private final Project project;
-	private final ProjectExportConfiguration configuration;
-
 	private static final String CONTROLS_BACKGROUND = "ControlsBackground";
 	private static final String OBJECTS = "Objects";
 	private static final String CONTROLS = "Controls";
-
-	private ResourceBundle bundle = Lang.ApplicationBundle();
-
-	public ProjectExporter(@NotNull ProjectExportConfiguration configuration) {
-		if (true) {
-			//Hey you. Convert this class to use util.IndentedStringBuilder
-			throw new RuntimeException();
-		}
-		this.configuration = configuration;
-		this.project = configuration.getProject();
-	}
 
 	public static void export(@NotNull ProjectExportConfiguration configuration) throws IOException {
 		new ProjectExporter(configuration).export();
@@ -47,12 +38,74 @@ public class ProjectExporter {
 		new ProjectExporter(configuration).export(displayOutputStream, macrosOutputStream);
 	}
 
+
+	@NotNull
+	public static String getMacrosFileName(@NotNull ProjectExportConfiguration configuration) {
+		return configuration.getExportClassName() + "_Macros" + configuration.getHeaderFileType().getExtension();
+	}
+
+	@NotNull
+	public static String getDisplayFileName(@NotNull ProjectExportConfiguration configuration) {
+		return configuration.getExportClassName() + configuration.getHeaderFileType().getExtension();
+	}
+
+	@NotNull
+	public static String getExportValueString(@NotNull SerializableValue value, @NotNull PropertyType type) {
+		String[] arr = value.getAsStringArray();
+		StringBuilder ret = new StringBuilder();
+		String v;
+		for (int i = 0; i < arr.length; i++) {
+			v = arr[i];
+			for (int quoteIndex : type.getIndexesWithQuotes()) {
+				if (quoteIndex == i) {
+					v = "\"" + v + "\"";
+					break;
+				}
+			}
+			ret.append(v).append(i != arr.length - 1 ? "," : "");
+		}
+		return arr.length > 1 ? "{" + ret + "}" : ret.toString();
+	}
+
+	public static void exportControlClass(@NotNull ProjectExportConfiguration configuration, @NotNull ControlClass controlClass, @NotNull OutputStream stream) throws IOException {
+		ProjectExporter exporter = new ProjectExporter(configuration);
+		CachedIndentedStringBuilder builder = getBuilder(stream);
+		exporter.writeControlClass(builder, controlClass);
+		stream.flush();
+	}
+
+	private static CachedIndentedStringBuilder getBuilder(@NotNull OutputStream outputStream) {
+		return new CachedIndentedStringBuilder(4, true, 120, s -> {
+			try {
+				outputStream.write(s.getBytes());
+				outputStream.flush();
+			} catch (IOException e) {
+				throw new IllegalStateException(e);
+			}
+			return null;
+		});
+	}
+
+	private final Project project;
+	private final ProjectExportConfiguration configuration;
+	private final ResourceBundle bundle = Lang.ApplicationBundle();
+
+	private CachedIndentedStringBuilder displayStringBuilder;
+
+	private CachedIndentedStringBuilder macrosStringBuilder;
+
+	public ProjectExporter(@NotNull ProjectExportConfiguration configuration) {
+		this.configuration = configuration;
+		this.project = configuration.getProject();
+
+	}
+
 	public void export() throws IOException {
 		if (!configuration.getExportLocation().exists()) {
 			configuration.getExportLocation().mkdirs();
 		}
 		if (!configuration.getExportLocation().isDirectory()) {
-			throw new IllegalArgumentException("exportLocation ('"+configuration.getExportLocation().getPath()+"') is not a directory");
+			throw new IllegalArgumentException("exportLocation ('" + configuration.getExportLocation().getPath() + "') is not a directory");
 		}
 		final String exportDirectoryPath = configuration.getExportLocation().getPath() + "/";
 		final File exportDirectory = configuration.getExportLocation();
@@ -61,13 +114,13 @@ public class ProjectExporter {
 
 		final File exportDisplayFile = new File(exportDirectoryPath + getDisplayFileName(configuration));
 		exportDisplayFile.createNewFile();
-		final FileOutputStream fosDisplay = getFos(exportDisplayFile);
+		final FileOutputStream fosDisplay = new FileOutputStream(exportDisplayFile);
 
 		FileOutputStream fosMacros = null;
 		if (configuration.shouldExportMacrosToFile()) {
 			final File macrosExportFile = new File(exportDirectoryPath + getMacrosFileName(configuration));
 			macrosExportFile.createNewFile();
-			fosMacros = getFos(macrosExportFile);
+			fosMacros = new FileOutputStream(macrosExportFile);
 		}
 
 		export(fosDisplay, fosMacros);
@@ -80,111 +133,148 @@ public class ProjectExporter {
 
 	public void export(@NotNull OutputStream displayOutputStream, @Nullable OutputStream macrosOutputStream) throws IOException {
 		if (macrosOutputStream == null || !configuration.shouldExportMacrosToFile()) {
-			exportMacros(displayOutputStream); //save the macros inside the display header file
-		} else {
-			exportMacros(macrosOutputStream);
-			if (configuration.shouldPlaceAdcNotice()) {
-				writelnComment(displayOutputStream, bundle.getString("Misc.adc_export_notice"));
-				writeln(displayOutputStream, "");
-			}
+			macrosOutputStream = displayOutputStream; //save the macros inside the display header file
 		}
 
-		exportDisplay(displayOutputStream);
+		displayStringBuilder = getBuilder(displayOutputStream);
+		macrosStringBuilder = getBuilder(macrosOutputStream);
 
-		displayOutputStream.flush();
-		if (macrosOutputStream != null && macrosOutputStream != displayOutputStream) {
-			macrosOutputStream.flush();
-		}
-	}
+		exportMacros(macrosStringBuilder);
+		//write remainder stuff
+		macrosOutputStream.write(macrosStringBuilder.toString().getBytes());
 
-	public static void exportControlClass(@NotNull ProjectExportConfiguration configuration, @NotNull ControlClass controlClass, @NotNull OutputStream stream) throws IOException {
-		ProjectExporter exporter = new ProjectExporter(configuration);
-		stream.write(exporter.getExportClassString(controlClass, 0, null).getBytes());
-		stream.flush();
-	}
-
-
-	private void exportMacros(@NotNull OutputStream os) throws IOException {
 		if (configuration.shouldPlaceAdcNotice()) {
-			writelnComment(os, bundle.getString("Misc.adc_export_notice"));
-			writeln(os, "");
+			writelnComment(displayStringBuilder, bundle.getString("Misc.adc_export_notice"));
+			writeln(displayStringBuilder, "");
+		}
+
+		exportDisplay(displayStringBuilder);
+
+		//write the remainder of the string builders
+		displayOutputStream.write(displayStringBuilder.toString().getBytes());
+
+		//one last flush
+		displayOutputStream.flush();
+		macrosOutputStream.flush();
+	}
+
+	private void writeln(@NotNull IndentedStringBuilder stringBuilder, @NotNull String s) {
+		stringBuilder.append(s);
+		stringBuilder.append('\n');
+	}
+
+
+	private void exportMacros(@NotNull IndentedStringBuilder stringBuilder) throws IOException {
+		if (configuration.shouldPlaceAdcNotice()) {
+			writelnComment(stringBuilder, bundle.getString("Misc.adc_export_notice"));
+			writeln(stringBuilder, "");
 		}
 
 		for (Macro macro : project.getMacroRegistry().getMacros()) {
 			if (macro.getComment() != null && macro.getComment().length() != 0) {
-				writelnComment(os, macro.getComment());
+				writelnComment(stringBuilder, macro.getComment());
 			}
-			writeln(os, "#define " + macro.getKey() + " " + getExportValueString(macro.getValue(), macro.getPropertyType()));
+			writeln(stringBuilder, "#define " + macro.getKey() + " " + getExportValueString(macro.getValue(), macro.getPropertyType()));
 		}
 
-		writeln(os, "");
+		stringBuilder.append('\n');
 	}
 
-	private void exportDisplay(@NotNull OutputStream os) throws IOException {
+	private void exportDisplay(@NotNull IndentedStringBuilder stringBuilder) throws IOException {
 		if (configuration.shouldExportMacrosToFile()) {
-			writeln(os, "#include \"" + getMacrosFileName(configuration) + "\"");
-			writeln(os, "");
+			writeln(stringBuilder, "#include \"" + getMacrosFileName(configuration) + "\"");
+			writeln(stringBuilder, "");
 		}
 		if (project.getProjectDescription() != null && project.getProjectDescription().length() > 0) {
-			writelnComment(os, project.getProjectDescription());
+			writelnComment(stringBuilder, project.getProjectDescription());
 		}
 		ArmaDisplay display = project.getEditingDisplay();
 
-		String displayBody = getExportControlPropertyString(1, display.getDisplayProperties()) + "\n\n";
-		String body = "";
-		for (ArmaControl control : display.getBackgroundControls()) {
-			body += getExportControlString(control, 2);
-		}
-		displayBody += getExportClassString(CONTROLS_BACKGROUND, null, 1, body) + "\n";
+		writeClass(stringBuilder, configuration.getExportClassName(), null, stringBuilderCopy -> {
+			//write display properties
 
-		body = "";
-		for (ArmaControl control : display.getControls()) {
-			body += getExportControlString(control, 2);
-		}
-		displayBody += getExportClassString(CONTROLS, null, 1, body);
+			writeControlProperties(displayStringBuilder, display.getDisplayProperties());
+			displayStringBuilder.append('\n');
 
-		writeln(os, getExportClassString(configuration.getExportClassName(), null, 0, displayBody));
+
+			//write background controls
+			writeClass(stringBuilder, CONTROLS_BACKGROUND, null, sb -> {
+				for (ArmaControl control : display.getBackgroundControls()) {
+					writeControl(stringBuilder, control);
+				}
+				return null;
+			});
+
+			//write controls
+			writeClass(stringBuilder, CONTROLS, null, sb -> {
+				for (ArmaControl control : display.getControls()) {
+					writeControl(stringBuilder, control);
+				}
+				return null;
+			});
+
+
+			return null;
+		});
 
 	}
 
-	private String getExportControlString(@NotNull ArmaControl control, int tab) {
+	private void writeControl(@NotNull IndentedStringBuilder stringBuilder, @NotNull ArmaControl control) {
 		if (control instanceof ArmaControlGroup) {
-			String controlsBody = "";
+			//write group's "Controls" class
 			for (ArmaControl subControl : ((ArmaControlGroup) control).getControls()) {
-				controlsBody += getExportControlString(subControl, tab + 2);
+				writeClass(stringBuilder, CONTROLS, null, sb -> {
+					writeControl(sb, subControl);
+					return null;
+				});
 			}
-			return getExportClassString(control, tab, getExportClassString(CONTROLS, null, tab + 1, controlsBody));
+
 		}
-		return getExportClassString(control, tab, null);
+		//write control body
+		writeControlClass(stringBuilder, control);
 	}
 
-	private String getExportClassString(@NotNull ControlClass controlClass, int tab, @Nullable String additionalBodyContent) {
-		String body = getExportControlPropertyString(tab + 1, controlClass.getDefinedProperties());
-
-		String tabS = tab(tab);
-		for (ControlClass subclass : controlClass.getAllNestedClasses()) {
-			body += tabS + getExportClassString(subclass, tab + 1, null);
-		}
-		if (additionalBodyContent != null) {
-			body += "\n";
-			if (!additionalBodyContent.startsWith(tabS)) {
-				body += tabS;
+	private void writeControlClass(@NotNull IndentedStringBuilder stringBuilder, @NotNull ControlClass controlClass) {
+		writeClass(stringBuilder, controlClass.getClassName(), controlClass.getExtendClass() == null ? null : controlClass.getExtendClass().getClassName(), sb -> {
+			writeControlProperties(sb, controlClass.getAllChildProperties());
+			for (ControlClass nested : controlClass.getAllNestedClasses()) {
+				writeControlClass(stringBuilder, nested);
 			}
-			body += additionalBodyContent;
+			return null;
+		});
+	}
+
+	private void writeClass(@NotNull IndentedStringBuilder stringBuilder, @NotNull String className, @Nullable String extendClassName, @NotNull Function<IndentedStringBuilder, Void> writeBodyFunc) {
+		//class example : thing
+		//{
+		//	writeBodyFunc.apply(stringBuilder)
+		//};
+
+		stringBuilder.append("class ");
+		stringBuilder.append(className);
+		if (extendClassName != null) {
+			stringBuilder.append(" : ");
+			stringBuilder.append(extendClassName);
+			stringBuilder.append(" ");
 		}
-		return getExportClassString(controlClass.getClassName(), controlClass.getExtendClass() != null ? controlClass.getExtendClass().getClassName() : null, tab, body);
+
+		stringBuilder.append('\n');
+		stringBuilder.incrementTabCount();
+		stringBuilder.append('{');
+
+		stringBuilder.append('\n');
+
+		writeBodyFunc.apply(stringBuilder);
+
+		stringBuilder.decrementTabCount();
+		stringBuilder.append('\n');
+		stringBuilder.append("};");
+		stringBuilder.append('\n');
 	}
 
-	private String getExportClassString(@NotNull String className, @Nullable String extendClass, int parentTab, String body) {
-		final String classFormatString = "%3$sclass %s%s\n%3$s{\n%4$s\n%3$s};";
-		return String.format(classFormatString, className, extendClass != null ? (" : " + extendClass) : "", tab(parentTab), body);
-	}
-
-	private static String getExportControlPropertyString(int tabNum, @NotNull Iterable<? extends ControlProperty> controlProperties) {
-		String body = "";
-		final String tab = tab(tabNum);
-		final String itemFormatString = tab + "%s = %s;";
-		final String itemArrayFormatString = tab + "%s[] = %s;";
+	private void writeControlProperties(@NotNull IndentedStringBuilder stringBuilder, @NotNull Iterable<? extends ControlProperty> controlProperties) {
+		String itemFormatString = "%s = %s;";
+		String itemArrayFormatString = "%s[] = %s;";
 		for (ControlProperty property : controlProperties) {
 			if (property.getValue() == null/* && editor.isOptional()*/) { //can allow for partial implementation, so we don't need to check if it is optional
 				continue;
@@ -197,73 +287,49 @@ public class ProjectExporter {
 				if (property.getMacro() instanceof StringTableKey) {
 					stringKey = (StringTableKey) property.getMacro();
 				}
-				body += String.format(itemFormatString, property.getName(), stringKey != null ? stringKey.getHeaderMacroId() : property.getMacro().getKey());
+				stringBuilder.append(String.format(itemFormatString, property.getName(), stringKey != null ? stringKey.getHeaderMacroId() : property.getMacro().getKey()));
 			} else {
 				if (property.getValue().getAsStringArray().length == 1) {
-					body += String.format(itemFormatString, property.getName(), getExportValueString(property.getValue(), property.getPropertyType()));
+					stringBuilder.append(String.format(itemFormatString, property.getName(), getExportValueString(property.getValue(), property.getPropertyType())));
 				} else {
-					body += String.format(itemArrayFormatString, property.getName(), getExportValueString(property.getValue(), property.getPropertyType()));
+					stringBuilder.append(String.format(itemArrayFormatString, property.getName(), getExportValueString(property.getValue(), property.getPropertyType())));
 				}
 			}
-			body += "\n";
+			stringBuilder.append('\n');
 		}
-		return body.substring(0, body.length() - 1); //exclude the last newline character
 	}
 
-	private static String tab(int num) {
-		String s = "";
-		for (int t = 0; t < num; t++) {
-			s += "\t";
-		}
-		return s;
-	}
-
-	@NotNull
-	public static String getMacrosFileName(@NotNull ProjectExportConfiguration configuration) {
-		return configuration.getExportClassName() + "_Macros" + configuration.getHeaderFileType().getExtension();
-	}
-
-	@NotNull
-	public static String getDisplayFileName(@NotNull ProjectExportConfiguration configuration) {
-		return configuration.getExportClassName() + configuration.getHeaderFileType().getExtension();
-	}
-
-	public static String getExportValueString(@NotNull SerializableValue value, @NotNull PropertyType type) {
-		String[] arr = value.getAsStringArray();
-		String ret = "";
-		String v;
-		for (int i = 0; i < arr.length; i++) {
-			v = arr[i];
-			for (int quoteIndex : type.getIndexesWithQuotes()) {
-				if (quoteIndex == i) {
-					v = "\"" + v + "\"";
-					break;
-				}
-			}
-			ret += v + (i != arr.length - 1 ? "," : "");
-		}
-		return arr.length > 1 ? "{" + ret + "}" : ret;
-	}
-
-	private static void writelnComment(OutputStream os, String s) throws IOException {
+	private void writelnComment(@NotNull IndentedStringBuilder stringBuilder, String s) throws IOException {
 		if (s.contains("\n")) {
-			writeln(os, "/*");
-			writeln(os, s);
-			writeln(os, "*/");
+			writeln(stringBuilder, "/*");
+			writeln(stringBuilder, s);
+			writeln(stringBuilder, "*/");
 		} else {
-			writeln(os, "//" + s);
+			writeln(stringBuilder, "//" + s);
 		}
 	}
 
-	private static void writeln(OutputStream fos, String s) throws IOException {
-		fos.write((s + "\n").getBytes());
-	}
+	private static class CachedIndentedStringBuilder extends IndentedStringBuilder {
 
-	private static FileOutputStream getFos(@NotNull File file) {
-		try {
-			return new FileOutputStream(file);
-		} catch (FileNotFoundException e) {
-			throw new IllegalArgumentException(e);
+		private final int cacheSize;
+		private int appendCount;
+		private Function<String, Object> onFull;
+
+		public CachedIndentedStringBuilder(int tabSizeInSpaces, boolean useTabCharacter, int cacheSize, @NotNull Function<String, Object> onFull) {
+			super(tabSizeInSpaces, useTabCharacter);
+			this.cacheSize = cacheSize;
+			this.onFull = onFull;
+		}
+
+		@Override
+		public void append(char c) {
+			super.append(c);
+			appendCount++;
+			if (appendCount >= cacheSize) {
+				onFull.apply(this.toString());
+				this.getBuilder().delete(0, length());
+				appendCount = 0;
+			}
 		}
 	}
 }